Configuraciones iniciales: Git ignore

+++ src/main/java/xyz/septimocielo/dao/VideojuegoDAO.java
@@ -3,6 +3,6 @@ package xyz.septimocielo.dao;
 import org.springframework.data.jpa.repository.JpaRepository;
 import xyz.septimocielo.model.Videojuego;
 
-public interface VideojuegoDAO extends JpaRepository<Videojuego, Integer> {
+public interface VideojuegoDAO extends JpaRepository<Videojuego, Long> {
 
 }
